refactor: 검증 로직  및 에러메세지 처리 수정
